+++ crypto/src/commonMain/kotlin/at/asitplus/awesn1/crypto/pki/Pkcs10CertificationRequest.kt
@@ -3,8 +3,10 @@
 
 package at.asitplus.awesn1.crypto.pki
 
-import at.asitplus.awesn1.crypto.X509AlgorithmIdentifier
+import at.asitplus.awesn1.WithPemLabel
+import at.asitplus.awesn1.WithValidPemLabels
 import at.asitplus.awesn1.crypto.SignatureValue
+import at.asitplus.awesn1.crypto.X509AlgorithmIdentifier
 import kotlinx.serialization.Serializable
 
 /**
@@ -23,9 +25,13 @@ data class Pkcs10CertificationRequest(
     val certificationRequestInfo: Pkcs10CertificationRequestInfo,
     val signatureAlgorithm: X509AlgorithmIdentifier,
     val signatureValue: SignatureValue,
-) {
-    companion object {
-        //NOT PEM DECODCABLE/ENCODABLE to avoid infinite recursion.
-        const val PEM_LABEL = "CERTIFICATE REQUEST"
+) : WithPemLabel {
+    override val pemLabel: String get() = canonicalPemLabel
+
+    companion object : WithValidPemLabels<Pkcs10CertificationRequest> {
+
+        override val canonicalPemLabel: String get() = "CERTIFICATE REQUEST"
+
+        override val validPemLabels: Set<String> = setOf(canonicalPemLabel, "NEW CERTIFICATE REQUEST")
     }
 }