+++ crypto/src/commonMain/kotlin/at/asitplus/awesn1/crypto/pki/X509Certificate.kt
@@ -3,8 +3,10 @@
 
 package at.asitplus.awesn1.crypto.pki
 
-import at.asitplus.awesn1.crypto.X509AlgorithmIdentifier
+import at.asitplus.awesn1.WithPemLabel
+import at.asitplus.awesn1.WithValidPemLabels
 import at.asitplus.awesn1.crypto.SignatureValue
+import at.asitplus.awesn1.crypto.X509AlgorithmIdentifier
 import kotlinx.serialization.Serializable
 
 /**
@@ -23,8 +25,10 @@ data class X509Certificate(
     val tbsCertificate: X509TbsCertificate,
     val signatureAlgorithm: X509AlgorithmIdentifier,
     val signatureValue: SignatureValue,
-) {
-    companion object {
-        const val PEM_LABEL = "CERTIFICATE"
+) : WithPemLabel {
+    override val pemLabel: String get() = canonicalPemLabel
+
+    companion object : WithValidPemLabels<X509Certificate> {
+        override val canonicalPemLabel: String get() = "CERTIFICATE"
     }
 }