+++ crypto/src/commonTest/kotlin/at/asitplus/awesn1/crypto/legacy/pki/X509Certificate.kt
@@ -23,7 +23,7 @@ open class X509Certificate(
     val signatureValue: SignatureValue,
 ) : Asn1PemEncodable<Asn1Sequence> {
 
-    override val label get() = canonicalPemLabel
+    override val pemLabel get() = canonicalPemLabel
 
     override fun encodeToTlv() = Asn1.Sequence {
         +tbsCertificate